+++ packages/medusa-plugin-sendgrid/src/services/sendgrid.js
@@ -34,6 +34,7 @@ class SendGridService extends NotificationService {
       totalsService,
       productVariantService,
       giftCardService,
+      logger,
     },
     options
   ) {
@@ -53,6 +54,7 @@ class SendGridService extends NotificationService {
     this.totalsService_ = totalsService
     this.productVariantService_ = productVariantService
     this.giftCardService_ = giftCardService
+    this.logger_ = logger
 
     SendGrid.setApiKey(options.api_key)
   }
@@ -219,26 +221,26 @@ class SendGridService extends NotificationService {
   }
 
   async sendNotification(event, eventData, attachmentGenerator) {
+    const data = await this.fetchData(event, eventData, attachmentGenerator)
+
     let templateId = this.getTemplateId(event)
 
-    if (!templateId) {
-      throw new MedusaError(MedusaError.Types.INVALID_DATA, `Sendgrid service: No template was set for event: ${event}`)
+    if (data.locale) {
+      templateId = this.getLocalizedTemplateId(event, data.locale) || templateId
     }
 
-    const data = await this.fetchData(event, eventData, attachmentGenerator)
-    if (!data) {
-      throw new MedusaError(MedusaError.Types.INVALID_DATA, "Sendgrid service: Invalid event data was received")
-   }
+    if (!templateId) {
+      throw new MedusaError(
+        MedusaError.Types.INVALID_DATA,
+        `Sendgrid service: No template was set for event: ${event}`
+      )
+    }
 
     const attachments = await this.fetchAttachments(
       event,
       data,
       attachmentGenerator
     )
-
-    if (data.locale) {
-      templateId = this.getLocalizedTemplateId(event, data.locale) || templateId
-    }
 
     const sendOptions = {
       template_id: templateId,
@@ -261,10 +263,16 @@ class SendGridService extends NotificationService {
       })
     }
 
-   let status
-   await this.transporter_.sendMail(sendOptions)
-   .then(() => { status = "sent" })
-   .catch((error) => { status = "failed"; console.log(error) })
+    let status
+
+    await SendGrid.send(sendOptions)
+      .then(() => {
+        status = "sent"
+      })
+      .catch((error) => {
+        status = "failed"
+        this.logger_.error(error)
+      })
 
     // We don't want heavy docs stored in DB
     delete sendOptions.attachments
@@ -303,18 +311,11 @@ class SendGridService extends NotificationService {
 
   /**
    * Sends an email using SendGrid.
-   * @param {string} templateId - id of template in SendGrid
-   * @param {string} from - sender of email
-   * @param {string} to - receiver of email
-   * @param {Object} data - data to send in mail (match with template)
+   * @param {Object} options - send options containing to, from, template, and more. Read more here: https://github.com/sendgrid/sendgrid-nodejs/tree/main/packages/mail
    * @return {Promise} result of the send operation
    */
   async sendEmail(options) {
-    try {
-      return SendGrid.send(options)
-    } catch (error) {
-      throw error
-    }
+    return await SendGrid.send(options)
   }
 
   async orderShipmentCreatedData({ id, fulfillment_id }, attachmentGenerator) {
@@ -613,11 +614,8 @@ class SendGridService extends NotificationService {
     // Fetch the return request
     const returnRequest = await this.returnService_.retrieve(return_id, {
       relations: [
-        "items",
-        "items.item",
         "items.item.tax_lines",
-        "items.item.variant",
-        "items.item.variant.product",
+        "items.item.variant.product.profiles",
         "shipping_method",
         "shipping_method.tax_lines",
         "shipping_method.shipping_option",
@@ -628,7 +626,9 @@ class SendGridService extends NotificationService {
       {
         id: returnRequest.items.map(({ item_id }) => item_id),
       },
-      { relations: ["tax_lines", "variant", "variant.product"] }
+      {
+        relations: ["tax_lines", "variant", "variant.product.profiles"],
+      }
     )
 
     // Fetch the order
@@ -774,7 +774,7 @@ class SendGridService extends NotificationService {
     })
 
     const cart = await this.cartService_.retrieve(swap.cart_id, {
-      relations: ["items", "items.variant", "items.variant.product"],
+      relations: ["items.variant.product.profiles"],
       select: [
         "total",
         "tax_total",
@@ -855,8 +855,7 @@ class SendGridService extends NotificationService {
     })
     const swap = await this.swapService_.retrieve(id, {
       relations: [
-        "additional_items",
-        "additional_items.variant.product",
+        "additional_items.variant.product.profiles",
         "additional_items.tax_lines",
         "return_order",
         "return_order.items",
@@ -873,7 +872,7 @@ class SendGridService extends NotificationService {
         id: returnRequest.items.map(({ item_id }) => item_id),
       },
       {
-        relations: ["tax_lines", "variant", "variant.product"],
+        relations: ["tax_lines", "variant.product.profiles"],
       }
     )
 
@@ -893,9 +892,7 @@ class SendGridService extends NotificationService {
     const order = await this.orderService_.retrieve(swap.order_id, {
       select: ["total"],
       relations: [
-        "items",
-        "items.variant",
-        "items.variant.product",
+        "items.variant.product.profiles",
         "items.tax_lines",
         "discounts",
         "discounts.rule",
@@ -915,7 +912,7 @@ class SendGridService extends NotificationService {
         "shipping_total",
         "subtotal",
       ],
-      relations: ["items", "items.variant", "items.variant.product"],
+      relations: ["items.variant.product.profiles"],
     })
     const currencyCode = order.currency_code.toUpperCase()
 
@@ -997,9 +994,7 @@ class SendGridService extends NotificationService {
         "shipping_methods",
         "shipping_methods.shipping_option",
         "shipping_methods.tax_lines",
-        "additional_items",
-        "additional_items.variant",
-        "additional_items.variant.product",
+        "additional_items.variant.product.profiles",
         "additional_items.tax_lines",
         "return_order",
         "return_order.items",
@@ -1011,14 +1006,11 @@ class SendGridService extends NotificationService {
         "region",
         "items",
         "items.tax_lines",
-        "items.variant",
-        "items.variant.product",
+        "items.variant.product.profiles",
         "discounts",
         "discounts.rule",
         "swaps",
-        "swaps.additional_items",
-        "swaps.additional_items.variant",
-        "swaps.additional_items.variant.product",
+        "swaps.additional_items.variant.product.profiles",
         "swaps.additional_items.tax_lines",
       ],
     })
@@ -1031,7 +1023,7 @@ class SendGridService extends NotificationService {
         "shipping_total",
         "subtotal",
       ],
-      relations: ["items", "items.variant", "items.variant.product"],
+      relations: ["items.variant.product.profiles"],
     })
 
     const returnRequest = swap.return_order
@@ -1040,7 +1032,7 @@ class SendGridService extends NotificationService {
         id: returnRequest.items.map(({ item_id }) => item_id),
       },
       {
-        relations: ["tax_lines", "variant", "variant.product"],
+        relations: ["tax_lines", "variant.product.profiles"],
       }
     )
 
@@ -1147,10 +1139,7 @@ class SendGridService extends NotificationService {
   async claimShipmentCreatedData({ id, fulfillment_id }) {
     const claim = await this.claimService_.retrieve(id, {
       relations: [
-        "order",
-        "order.items",
-        "order.items.variant",
-        "order.items.variant.product",
+        "order.items.variant.product.profiles",
         "order.shipping_address",
       ],
     })