+++ packages/medusa/src/workflows/admin/create-products/types.ts
@@ -1,0 +1,10 @@
+import { EntityManager } from "typeorm"
+import { MedusaContainer } from "@medusajs/types"
+import { AdminPostProductsReq } from "../../../api"
+
+export type InjectedDependencies = {
+  manager: EntityManager
+  container: MedusaContainer
+}
+
+export type CreateProductsWorkflowInputData = AdminPostProductsReq[]