+++ packages/medusa/src/workflows/functions/remove-products.ts
@@ -1,0 +1,13 @@
+import { MedusaContainer, ProductTypes } from "@medusajs/types"
+
+export async function removeProducts({
+  container,
+  data,
+}: {
+  container: MedusaContainer
+  data: ProductTypes.ProductDTO[]
+}) {
+  const productModuleService: ProductTypes.IProductModuleService =
+    container.resolve("productModuleService")
+  return await productModuleService.softDelete(data.map((p) => p.id))
+}