+++ packages/product/src/repositories/product-category.ts
@@ -1,35 +1,36 @@
-import { SqlEntityManager } from "@mikro-orm/postgresql"
 import {
   FilterQuery as MikroFilterQuery,
   FindOptions as MikroOptions,
   LoadStrategy,
 } from "@mikro-orm/core"
-import { deduplicateIfNecessary } from "../utils"
-import { ProductCategory } from "@models"
-import { DAL, ProductCategoryTransformOptions } from "@medusajs/types"
+import { Product, ProductCategory } from "@models"
+import { Context, DAL, ProductCategoryTransformOptions } from "@medusajs/types"
 import groupBy from "lodash/groupBy"
+import { AbstractTreeRepositoryBase } from "./base"
+import { SqlEntityManager } from "@mikro-orm/postgresql"
+import { InjectTransactionManager, MedusaContext } from "@medusajs/utils"
 
-export class ProductCategoryRepository
-  implements DAL.RepositoryService<ProductCategory>
-{
+export class ProductCategoryRepository extends AbstractTreeRepositoryBase<ProductCategory> {
   protected readonly manager_: SqlEntityManager
 
-  constructor({ manager }) {
-    this.manager_ = manager.fork()
+  constructor({ manager }: { manager: SqlEntityManager }) {
+    // @ts-ignore
+    super(...arguments)
+    this.manager_ = manager
   }
 
   async find(
     findOptions: DAL.FindOptions<ProductCategory> = { where: {} },
     transformOptions: ProductCategoryTransformOptions = {},
-    context: { transaction?: any } = {}
+    context: Context = {}
   ): Promise<ProductCategory[]> {
-    // Spread is used to copy the options in case of manipulation to prevent side effects
+    const manager = (context.transactionManager ??
+      this.manager_) as SqlEntityManager
+
     const findOptions_ = { ...findOptions }
     const { includeDescendantsTree } = transformOptions
-
     findOptions_.options ??= {}
     const fields = (findOptions_.options.fields ??= [])
-    findOptions_.options.limit ??= 15
 
     // Ref: Building descendants
     // mpath and parent_category_id needs to be added to the query for the tree building to be done accurately
@@ -39,19 +40,11 @@ export class ProductCategoryRepository
         fields.push("parent_category_id")
     }
 
-    if (findOptions_.options.populate) {
-      deduplicateIfNecessary(findOptions_.options.populate)
-    }
-
-    if (context.transaction) {
-      Object.assign(findOptions_.options, { ctx: context.transaction })
-    }
-
     Object.assign(findOptions_.options, {
       strategy: LoadStrategy.SELECT_IN,
     })
 
-    const productCategories = await this.manager_.find(
+    const productCategories = await manager.find(
       ProductCategory,
       findOptions_.where as MikroFilterQuery<ProductCategory>,
       findOptions_.options as MikroOptions<ProductCategory>
@@ -69,8 +62,12 @@ export class ProductCategoryRepository
 
   async buildProductCategoriesWithDescendants(
     productCategories: ProductCategory[],
-    findOptions: DAL.FindOptions<ProductCategory> = { where: {} }
+    findOptions: DAL.FindOptions<ProductCategory> = { where: {} },
+    context: Context = {}
   ): Promise<ProductCategory[]> {
+    const manager = (context.transactionManager ??
+      this.manager_) as SqlEntityManager
+
     for (let productCategory of productCategories) {
       const whereOptions = {
         ...findOptions.where,
@@ -78,9 +75,11 @@ export class ProductCategoryRepository
           $like: `${productCategory.mpath}%`,
         },
       }
-      delete whereOptions.parent_category_id
 
-      const descendantsForCategory = await this.manager_.find(
+      delete whereOptions.parent_category_id
+      delete whereOptions.id
+
+      const descendantsForCategory = await manager.find(
         ProductCategory,
         whereOptions as MikroFilterQuery<ProductCategory>,
         findOptions.options as MikroOptions<ProductCategory>
@@ -111,30 +110,64 @@ export class ProductCategoryRepository
   async findAndCount(
     findOptions: DAL.FindOptions<ProductCategory> = { where: {} },
     transformOptions: ProductCategoryTransformOptions = {},
-    context: { transaction?: any } = {}
+    context: Context = {}
   ): Promise<[ProductCategory[], number]> {
-    // Spread is used to copy the options in case of manipulation to prevent side effects
+    const manager = (context.transactionManager ??
+      this.manager_) as SqlEntityManager
+
     const findOptions_ = { ...findOptions }
-
+    const { includeDescendantsTree } = transformOptions
     findOptions_.options ??= {}
-    findOptions_.options.limit ??= 15
+    const fields = (findOptions_.options.fields ??= [])
 
-    if (findOptions_.options.populate) {
-      deduplicateIfNecessary(findOptions_.options.populate)
-    }
-
-    if (context.transaction) {
-      Object.assign(findOptions_.options, { ctx: context.transaction })
+    // Ref: Building descendants
+    // mpath and parent_category_id needs to be added to the query for the tree building to be done accurately
+    if (includeDescendantsTree) {
+      fields.indexOf("mpath") === -1 && fields.push("mpath")
+      fields.indexOf("parent_category_id") === -1 &&
+        fields.push("parent_category_id")
     }
 
     Object.assign(findOptions_.options, {
       strategy: LoadStrategy.SELECT_IN,
     })
 
-    return await this.manager_.findAndCount(
+    const [productCategories, count] = await manager.findAndCount(
       ProductCategory,
       findOptions_.where as MikroFilterQuery<ProductCategory>,
       findOptions_.options as MikroOptions<ProductCategory>
     )
+
+    if (!includeDescendantsTree) {
+      return [productCategories, count]
+    }
+
+    return [
+      await this.buildProductCategoriesWithDescendants(
+        productCategories,
+        findOptions_
+      ),
+      count,
+    ]
+  }
+
+  @InjectTransactionManager()
+  async delete(
+    ids: string[],
+    @MedusaContext()
+    { transactionManager: manager }: Context = {}
+  ): Promise<void> {
+    await (manager as SqlEntityManager).nativeDelete(
+      Product,
+      { id: { $in: ids } },
+      {}
+    )
+  }
+
+  async create(
+    data: unknown[],
+    context: Context = {}
+  ): Promise<ProductCategory[]> {
+    throw new Error("Method not implemented.")
   }
 }