+++ packages/types/src/shared-context.ts
@@ -3,3 +3,10 @@ import { EntityManager } from "typeorm"
 export type SharedContext = {
   transactionManager?: EntityManager
 }
+
+export type Context<TManager = unknown> = {
+  transactionManager?: TManager
+  isolationLevel?: string
+  enableNestedTransactions?: boolean
+  transactionId?: string
+}